add flashlight, bg, bgSound
