+++ Assets/Domain/Audio/timbre__ominous-bassy-rumble-drone.flac.meta
@@ -1,0 +1,22 @@
+fileFormatVersion: 2
+guid: e52f4ec7f655bec4cb37c99a8dd7777b
+AudioImporter:
+  externalObjects: {}
+  serializedVersion: 6
+  defaultSettings:
+    loadType: 0
+    sampleRateSetting: 0
+    sampleRateOverride: 44100
+    compressionFormat: 1
+    quality: 1
+    conversionMode: 0
+  platformSettingOverrides: {}
+  forceToMono: 0
+  normalize: 1
+  preloadAudioData: 1
+  loadInBackground: 0
+  ambisonic: 0
+  3D: 1
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
