+++ Assets/Domain/Scripts/Event/FlickeringLight.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: cdf58a5da028882499560a3d0a4e69aa
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
