+++ Assets/Domain/Scripts/Event/RingingEvent.cs
@@ -1,0 +1,33 @@
+//using System.Collections;
+//using System.Collections.Generic;
+//using Unity.VisualScripting;
+//using UnityEngine;
+
+//public class RingingEvent : MonoBehaviour
+//{
+//    //핸드폰
+//    [SerializeField]
+//    GameObject phone; //핸드폰에서 소리 재생
+
+//    [SerializeField]
+//    Transform tf; //플레이어 위치
+
+//    float dir;
+//    // Start is called before the first frame update
+//    void Start()
+//    {
+        
+//    }
+
+//// Update is called once per frame
+//    void Update()
+//    {
+//        dir = 
+//    }
+
+//    void OnCollisionEnter(Collision col)
+//    {
+//        if (col.gameObject.name == "handle")
+//            sound.PlayOneShot(sound.clip);
+//    }
+//}
