+++ Assets/Domain/Scripts/Event/RingingEvent.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 0864bb2a906f6864f839fc05774fe4ed
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
