+++ Assets/Domain/Scripts/Player/FlashLight.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 8dc59f3b5b815654194f4a9f40413291
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
